Add footer to the result page
This page is a special case where the body height changes.
To keep footer always at the bottom means it must "attach"
to the body but relatively stay at the bottom of the frame.

See this article for the trick in detail:
www.freecodecamp.org/news/how-to-keep-your-footer-where-it-belongs-59c6aa05c59c